Added relative imports support
Previous Goscriptify was copying source files to the temp directory, and
building from there. This caused issues with relative imports, as none
of the local *(to the source file)* files existed in the temp directory.

To solve this, Goscriptify now does one of three things for each source
file it's given.

1. Compiles the file directly, if it ends in .go already.
2. Copies the file to a .go extension of it's original name. Eg: `source
-> source.go` and uses that as the source. *(Removing it after)*
3. If `source` is given, but `source.go` already exists, a verbose hash
filename is used instead and used as source. Eg:
`source-sa8d78asd686adajkhadakjhsda.go`. This is simply a hash of the
current directory, and is the same hash used in all the source files
that this happens to, along with the binary compiled into the temp
directory.